requirements.txt
asgiref==3.2.7
Django==3.0.6
django-crispy-forms==1.9.1
django-environ==0.4.5
environ==1.0
mysqlclient==1.4.6
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1
